added notes from nam's prev original repo (his readme)
added nam's readme. going to decide to make it a note instead
of this repositories readme.

im thinking the readme should help first
time users of this repo to understand our project (general description
and motivation; install instruction; related links) this is just my initial
ideas and is subject to change

most of nam's readme were more of general notes for our own development
and benefit. i split it up to procastinate this issue of the readme for
later