init a base backend folder built of express generator
this branch will try to develop the backend using the express framework
of node js. The idea is that node js supports async which is good for
handling request in backend

using express may be temporary. other consideration of what to build the
backend of are with: Go, Rust, and Python.

Go and Rust seem to be high
demand in the current market and may be faster than js but difficulty is
in learning a new language.

Python technically requires less background
knowledge (cause school) but may need to learn async if not learned
already. I've seen some forums of supposed python dev complaining about
difficulty in learning async in python