basic implemention of chatbot with groq (not frontend implemented)
openai api is not free anymore i do not think. used groq api cause free
api. you need to sign up for api key from groq.

groq is prone to prompt injection; not secure. as of now you can ask it
to ignore the system message and it will :(

the syntax between groq and the openai are very similar; groq can be use
as a free temp until we want to actually specifically openai models.