updated package to reduce vuln, adjusted table creation, update readme
updated the readme to include the environment variable needed; since the
.env is no longer being stored in the repo, you must create the
environment variable yourself.

the environment variable mention is the database url to get the backend
connection to the database working. currently the backend will give
error if the database is not set up; consideration to adjust it so the
backend can work without.